fix import rnnrollout (#57)
* fix import rnnrollout

* fix import rnnrollout

* fix import, add visdom dependency

* add nltk download

* fix mdrg import

* fix import error
