+++ tatk/e2e/rnn_rollout/deal_or_not/__init__.py
@@ -1,0 +1,1 @@
+from tatk.e2e.rnn_rollout.deal_or_not.model import DealornotAgent
