+++ tatk/e2e/rnn_rollout/dialog.py
@@ -9,10 +9,10 @@ import pdb
 
 import numpy as np
 
-from tatk.e2e.rnn_rullout.metric import MetricsContainer
+from tatk.e2e.rnn_rollout.metric import MetricsContainer
 import data
-import tatk.e2e.rnn_rullout.utils as utils
-import tatk.e2e.rnn_rullout.domain as domain
+import tatk.e2e.rnn_rollout.utils as utils
+import tatk.e2e.rnn_rollout.domain as domain
 
 
 class DialogLogger(object):
