--- tatk/e2e/rnn_rullout/deal_or_not/__init__.py
@@ -1,1 +1,0 @@
-from tatk.e2e.rnn_rullout.deal_or_not.model import DealornotAgent
