+++ tatk/e2e/sequicity/metric.py
@@ -1,13 +1,11 @@
-# Modified by Microsoft Corporation.
-# Licensed under the MIT license.
-
 import argparse
 import csv
 import functools
 import json
 import math
 from collections import Counter
-
+import nltk
+nltk.download('stopwords')
 from nltk.corpus import stopwords
 from nltk.stem import WordNetLemmatizer
 from nltk.tokenize import word_tokenize
